src/chatui/requirements.txt
langchain
langchain_community
langchain_core
python-dotenv==1.0.1
streamlit==1.31.1
boto3==1.35.65
langchain_aws
psycopg2==2.9.10